autotimer/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
colorama==0.4.4
decorator==5.0.9
Deprecated==1.2.12
discord-ext-menus==1.1
discord.py==2.0.0a3391+gf14e584
idna==2.10
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
jaconv==0.3
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
matplotlib-inline==0.1.2
multidict==5.1.0
numpy==1.21.0
pandas==1.2.5
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.19
pycparser==2.20
Pygments==2.9.0
pykakasi==2.2.1
PyNaCl==1.4.0
python-dateutil==2.8.1
pytz==2021.1
pywin32==301
pyzmq==22.1.0
requests==2.25.1
six==1.16.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.5
wcwidth==0.2.5
wrapt==1.12.1
yarl==1.6.3